Add test server support for AppName and AppBasePath
