+++ src/Microsoft.AspNet.TestHost/TestServer.cs
@@ -4,6 +4,7 @@
 using System;
 using System.Collections.Generic;
 using System.Net.Http;
+using System.Runtime.Versioning;
 using System.Threading.Tasks;
 using Microsoft.AspNet.Builder;
 using Microsoft.AspNet.FeatureModel;
@@ -13,6 +14,7 @@ using Microsoft.AspNet.Hosting.Startup;
 using Microsoft.AspNet.Http;
 using Microsoft.Framework.ConfigurationModel;
 using Microsoft.Framework.DependencyInjection;
+using Microsoft.Framework.Runtime;
 using Microsoft.Framework.Runtime.Infrastructure;
 
 namespace Microsoft.AspNet.TestHost
@@ -21,6 +23,8 @@ namespace Microsoft.AspNet.TestHost
     {
         public IServiceProvider FallbackServices { get; set; }
         public string Environment { get; set; }
+        public string ApplicationName { get; set; }
+        public string ApplicationBasePath { get; set; }
 
         public Type StartupType { get; set; }
         public string StartupAssemblyName { get; set; }
@@ -37,6 +41,13 @@ namespace Microsoft.AspNet.TestHost
             if (Environment != null)
             {
                 config[HostingFactory.EnvironmentKey] = Environment;
+            }
+            if (ApplicationName != null || ApplicationBasePath != null)
+            {
+                var appEnv = new TestApplicationEnvironment(fallbackServices.GetRequiredService<IApplicationEnvironment>());
+                appEnv.ApplicationBasePath = ApplicationBasePath;
+                appEnv.ApplicationName = ApplicationName;
+                AdditionalServices.AddInstance<IApplicationEnvironment>(appEnv);
             }
 
             var engine = WebApplication.CreateHostingEngine(fallbackServices,
@@ -58,6 +69,67 @@ namespace Microsoft.AspNet.TestHost
 
             return new TestServer(engine);
         }
+
+        private class TestApplicationEnvironment : IApplicationEnvironment
+        {
+            private readonly IApplicationEnvironment _appEnv;
+            private string _appName;
+            private string _appBasePath;
+
+            public TestApplicationEnvironment(IApplicationEnvironment appEnv)
+            {
+                _appEnv = appEnv;
+            }
+
+            public string ApplicationBasePath
+            {
+                get
+                {
+                    return _appBasePath ?? _appEnv.ApplicationBasePath;
+                }
+                set
+                {
+                    _appBasePath = value;
+                }
+            }
+
+            public string ApplicationName
+            {
+                get
+                {
+                    return _appName ?? _appEnv.ApplicationName;
+                }
+                set
+                {
+                    _appName = value;
+                }
+            }
+
+            public string Configuration
+            {
+                get
+                {
+                    return _appEnv.Configuration;
+                }
+            }
+
+            public FrameworkName RuntimeFramework
+            {
+                get
+                {
+                    return _appEnv.RuntimeFramework;
+                }
+            }
+
+            public string Version
+            {
+                get
+                {
+                    throw new NotImplementedException();
+                }
+            }
+        }
+
     }
 
     public class TestServer : IServerFactory, IDisposable
